Feature/hu multinivel (#3)
* Se agregan y ajustan entidades para HU-Multinivel

* Se agregan y ajustan entidades para HU-Multinivel

* Se agregan y ajustan clases, utilidades. Adición paquetes para complementar estructura del proyecto, entre otros

* Se agregan y ajustan clases, utilidades. Adición paquetes para complementar estructura del proyecto, entre otros
