+++ administracion/src/main/java/com/administracion/constant/SQLConstant.java
@@ -1,0 +1,9 @@
+package com.administracion.constant;
+
+/**
+ * Clase constante que contiene los DMLs Y DDLs para las consultas nativas
+ */
+public class SQLConstant {
+	
+
+}
